+++ packages/starknet-snap/src/utils/din.test.ts
@@ -1,0 +1,23 @@
+import { constants } from 'starknet';
+
+import { getRPCUrl } from './din';
+
+describe('getRPCUrl', () => {
+  it('returns Mainnet RPC URL if chain id is Mainnet', () => {
+    expect(getRPCUrl(constants.StarknetChainId.SN_MAIN)).toBe(
+      'https://starknet-mainnet.infura.io/v3/',
+    );
+  });
+
+  it('returns Sepolia RPC URL if chain id is not either Mainnet or Sepolia', () => {
+    expect(getRPCUrl('0x534e5f474f45524c49')).toBe(
+      'https://starknet-sepolia.infura.io/v3/',
+    );
+  });
+
+  it('returns Sepolia RPC URL if chain id is Sepolia', () => {
+    expect(getRPCUrl(constants.StarknetChainId.SN_SEPOLIA)).toBe(
+      'https://starknet-sepolia.infura.io/v3/',
+    );
+  });
+});
